优化网络过滤，主动干掉连接，使网络阻止立即生效

+++ cmonitor/client/reports/hijack/HijackConfig.cs
@@ -84,11 +84,14 @@ namespace cmonitor.client.reports.hijack
                         DomainIPs[item] = false;
                 }
             }
-            List<int> adapters = Wininet.GetAdaptersIndex();
-            List<IPAddress> ips = DomainIPs.Where(c => c.Value == false).Select(c=>c.Key).ToList();
-            if(ips.Any() && adapters.Any())
+            if (OperatingSystem.IsWindows())
             {
-                Wininet.DeleteConnection(adapters, ips);
+                List<int> adapters = Wininet.GetAdaptersIndex();
+                List<IPAddress> ips = DomainIPs.Where(c => c.Value == false).Select(c => c.Key).ToList();
+                if (ips.Any() && adapters.Any())
+                {
+                    Wininet.DeleteConnection(adapters, ips);
+                }
             }
         }
